Fix modint constructor for negative inputs

+++ lib/math/modint.hpp
@@ -25,7 +25,7 @@ class modint {
     constexpr void set(const T &_x, bool raw = false) {
         if (raw || (0 <= _x && static_cast<uint32_t>(_x) < md)) x = _x;
         else {
-            long long tmp = _x % md;
+            long long tmp = _x % static_cast<int>(md);
             if (tmp < 0) tmp += md;
             x = static_cast<uint32_t>(tmp);
         }
